refactor: add data module
