+++ .eslintrc.json
@@ -11,5 +11,6 @@
         "sourceType": "module"
     },
     "rules": {
+        "import/extensions": 0
     }
 }
